Always do cleanup after a failed stack update
Move the cleanup code (that sets the status to FAILED and merges the new
environment into the old one) so that it runs on any exception, including
exit exceptions (due to e.g. the thread being cancelled or the engine
restarted). In the case of the environment this isn't as good as merging
them before the update starts (since it's possible for Heat to die without
time to run exception handlers), but at least it leaves things consistent
in the orderly case.

Change-Id: Ia4d2deca393ac9e250932c5cdb3691b6ee2b3ef4
Partial-bug: #1544348
Related-bug: #1492433