Make backup stack naming more consistent
We used to create the Stack object for the backup stack with the same name
as the actual stack, but store it in the database with an asterisk appended
to the name. This meant that a backup stack loaded from the database would
have an asterisk at the end of its name (e.g. in log messages), but one
newly created would not. Resolve this inconsistency by appending the
asterisk to the name at the time the Stack object is created, and don't
change it when we store it in the database.

Change-Id: I4eafddc7555fb148b160bdbe5a10ef649d4e27f2