Make parameter decryption more robust
Two bugs (#1572294 and #1521925) were reported where the "too many
values to unpack" exception was encountered. In both cases, certain
Heat functionality was broken.

Try to recover from this exception if it is encountered. If another
special case was missed, a warning will be logged so that the bug will
be seen and fixed.

Change-Id: I2cc0b262a89b95ba95945bd846171ba5270b802b