on-delete cascade rule for resource_data
During deleting of resource, the resource_data is also loaded in memory
and then deleted. This is not performant. The passive_delete option in
relationship deletes the relation without loading the children. This
patch configures the resource to have a passive-delete relationship with
resource_data and configures ondelete CASCADE so that the DB takes care
of deleting the resource_data when a resource is deleted.

This patch proposes the following changes:
- When deleting a resource do not load resource_data.
- When a resource is deleted, DB takes care of deleting resource_data.

The resource data remains intact when the FK constraints are deleted and
added to reflect on-delete cascade.

Change-Id: I4162e6f9ad9582fcb304887e5d3718de220231ec
Closes-Bug: #1573585