Revert "Replace SD RPC polling by long RPC call"
This partially reverts commit a709637a17166cb12e5d94944837bac6451b2o851.

Using long-polling causes a regression for non-convergence stacks, since
no other resources will be polled, started, or otherwise progressed
while we're waiting for the response. It can effectively serialise part
of the stack.

This patch leaves the new RPC call in place, since reverting that would
be a backward-incompatible change to the RPC API, and since we may find
a use for it once everyone is switched to convergence.

Change-Id: I14238618b5ec27c0e380383c2928ac56e605495f