Copy devstack code to heat tree
This begins work toward moving the devstack code to a devstack plugin within
heat tree. This allows us to experiment with the devstack code more easily,
as well as take some load off of the devstack team.
Note that the plugin is not enabled until we enable it in project-config,
so this doesn't actually affect devstack runs yet. The files that exist in
devstack are listed below.

$ find . -name *heat*
./lib/heat
./files/apache-heat-api-cfn.template
./files/debs/heat
./files/apache-heat-pip-repo.template
./files/apache-heat-api-cloudwatch.template
./files/apache-heat-api.template

All of these files are copied to heat tree.

Change-Id: Ie7f16eff554e6cc2c62823cd49b6f726d96f25b4