Allow scheduler tasks to request longer wait periods
This change allows a scheduler co-routine to yield an integer n, such
that only every nth call to TaskRunner.step() results in the co-routine
being advanced. We'll be able to use this to selectively slow down
polling of resources where and when we know it doesn't make sense to
poll every second.

Change-Id: Ifc6303524be8e477905546760279027886833617
Related-Bug: #1549219