Avoid full stack load for identify_stack
identify_stack is used to redirect API requests to the canonical stack
URL. The current approach results in 3 SQL queries to fetch the stack,
the raw template, and the tags.

This change uses the stack object directly to return the identifier.
It doesn't actually change any data access patterns but makes it
possible to avoid the loading of raw template and stack tags in future
changes.

Change-Id: Ic96e1492880368b3d25579893c10c5bd0ca2f165
Partial-Bug: #1578851