Lazy-load stack tags
Tags for a stack need to be loaded in the following scenarios:
- To display during a stack show
- To store during a create or update

All other operations which require a loaded stack will likely not
reference the tags, so the extra SQL query which always loads them is
often not necessary.

This change removes the tags loading from stack_object.Stack and
stack.Stack.load and implements annotated properties functions for a
Stack.tags property so that they are loaded from the database on
command.

Change-Id: Iecbf4f101b7f264a42b57f720a8428b0aed706e2
Partial-Bug: #1578851