Remove html_static_path from api-ref
In Id3a71841f5b9a9ea7bb5803afd967ab4feadfe10 the in tree copy of
os-api-ref was migrated to the upstream version. The html_static_path
setting in conf.py was forgotten in the patch, which caused a warning
when building the API reference.
This patch comments out the setting from conf.py to prevent the
warning.

TrivialFix

Change-Id: I267b3f671fdfa2033c24d9407c7bed7e8527e9a2