Don't use stored context to reset stacks
The resetting of in-progress stacks at startup is done using an admin
context. No user actions (e.g. ReST API calls) are performed, so there is
no need to load the user context that is stored with the stack. (Nor is
the
current context rewritten to the database in the process of resetting the
state.)

Since loading the user context from the DB may result in a call to the
keystone API, eliminating this ensures that stacks can be reset
successfully even if heat-engine starts before keystone is available.

Also there is no need to resolve outputs data while reset stacks, so set
resolve_data=False when load stack.

Change-Id: I9379ef6cc6f1d3bbc2569ff2150795bfaffee430
Closes-Bug: #1570569