Correct redundancies in Scheduler Hints docs
On the Heat Stack Lifecycle Scheduler Hints page [1], there are two
occurrences of redundant language. In the first and third paragraphs it
says "...passed *to nova and cinder* by heat as scheduler hints, *to the
configured schedulers for nova and cinder*."

"to the configured schedulers for nova and cinder" is removed from both
sentences in this patch.

[1]
http://docs.openstack.org/developer/heat/developing_guides/schedulerhints.html

Change-Id: I6f525986c82f850f86a20a232b96a36ed7d622cc
Closes-Bug: #1585792