Add period property to Monasca Notification resource
Currently, when one of Monasca's alarm transitions to the ALARM state, the
webhook associated with the Alarm is only invoked once. This is inconvenient
when the user wants to create a stack which will automatically scale up
or scale down more than once if the alarm state continues to be in the same
state.

The new added property 'PERIOD' will now allow the user to tell Monasca to
periodically invoke the webhook until the ALARM state transitions back to an OK
state.

To conform to the existing Heat autoscaling behaviour, we manually create the
monasca notification resource in Heat with a default value of 60 only for the
webhook notification type.

Change-Id: I2466911efa8d05dfe497f521885464ab90b0ae8e