Create an admin context to refresh template files
Its impractical to pass in the current context to
TemplateFiles._refresh, and it will soon be mandatory for all db api
context arguments to be not None. This change creates a short-lived
context just for the refresh.

It will have the same effect as the current behaviour of creating a
new session, since the a new session will be associated with the
context, so there is no danger of returning stale data from the
current session's identity map.

This change also makes the context argument mandatory for any raw
templates operations that might need one.

Change-Id: I156096d736aac4999a3eebe077ed50ef7384ca02