sqlalchemy tests don't assume same session
These tests currently rely on existing objects magically updating with
new values after an action which results in an update query. This
change refetches objects in preperation for the fix to bug #1479723.
This is closer in behaviour to how we actually use the db objects in
other parts of the codebase.

Change-Id: I303b3b20bd2b82cfeeac771cbb906ae2e6266b17
Related-Bug: #1479723