Always call TaskRunner._sleep()
In the patch 5f584db8d55a6db76c4307b21cb658a2bacb0296, we avoid calling
TaskRunner._sleep() after the first step if the wait_time is 0 or None.
This is wrong in the case of a wait_time of 0 - since this is distinct from
None, and should result in us calling eventlet.sleep(0). It's also
inconsistent in the case of wait_time=None, since for the rest of the steps
we call self._sleep(None) (and no actual sleep occurs).

This patch ensures we always call self._sleep(), passing 0 or None as
appropriate.

It also enhances the change in d66d57f187c7d8623526e9f928ae76d2cca9ecf9 to
ensure that we never yield from DependencyTaskGroup after all tasks are
completed, which could happen if that last batch of tasks to be started all
completed in a single step. This both saves changing the unit test to
accomodate the new _sleep(None) call and avoids an unnecessary
eventlet.sleep(0) call in real usage.

Change-Id: I301930fcba170e36d2d902c8d693ca722476cd18