Careful with OS::Nova::Server dependencies on subnets
A server within a template may depend on networks and, implicitly,
subnets, created within the same template. When we are calculating
dependencies for the server, play it safe and add any subnets as
dependencies whose network is not yet known. Note, this defensiveness
only takes places when subnets specify their network by reference,
rather by network id or name.

Change-Id: Ifda0a70b40e6f255b9ee462e8af7f11a0b2d5d6c
Closes-Bug: #1546201