Deprecate combination alarm
The combination alarm is deprecated and disabled
by default in Aodh, and will removed after two release
cycles in Aodh. Keep the same with Aodh, this change
deprecates combination alarm resource, before hidden it
we use ceilometer client as before because aodh client doesn't
support to manage combination alarm.

Blueprint migrate-to-use-aodh-for-alarms

Change-Id: Ibe8fe35a0cf9efe3d2809041ee480c99a75166cd