Remove unnecessary checks
Service's _get_stack method already checks
whether the stack is None and if it is, throws
an exception. So methods that use _get_stack
do not need to check the same thing.
Also, looks like there is no way that stack's
self.raw_template.template or self.raw_template.template
can be None or blank, so removed such checks as well
with corresponding tests.

Change-Id: I83d6e9251e0f7b6cacb0ddcfe1a3bc5dcf57d1c3