Fix empty values validation in nova server
In some places in validation code we assume '' value as
undefined. Howewer, '' means that property is assigned, but
not yet resolved, i.e. get_attr or get_resource function is
used and resolved to None and next in the code, when we get the value
of specific property, we convert None to default value. In this case
to empty string.

Closes-Bug: #1591098

Change-Id: I03a609ce87944c8efc9ec76a7a314d2de27d16d7