Do not raise NotFound for resource_get_all_* queries
A stack with zero resources is not an exception worthy condition (for
example, a stack being deleted, a stack which really does have zero
resources).

Also its unusual that a collection query function would raise a
NotFound when the result count is zero.

This change fixes that for resource_get_all_by_stack and
resource_get_all_active_by_stack

Change-Id: I3420d5099830f5626095b17c01e9d886b477915c
Related-Bug: #1301320
Related-Bug: #1523748