Fallback to db for resource show
This change adds a Stack.resource_get which is used by
describe_stack_resource (stack resource show). It has a different
behaviour when the resource is missing from the stack resources dict,
attempting to return a resource from the DB record in the current
stack.

This allows "stack resource show" to show a resource in a convergence
stack while the resource is being deleted.

Change-Id: I975778ac36ae31fe5bd3ebdbcd7c48a6fcc60757
Partial-Bug: #1301320
Related-Bug: #1523748