Cancel nested stack creation when sibling fails
When a resource fails and a sibling resource in the same stack is a nested
stack and is being created, we should (gracefully) cancel the creation of
the nested stack. This allows the user to correct the issue and try to
recover the stack with a stack update within a reasonable period of time.
Otherwise, they have to wait for the entire nested stack to complete, fail
or time out before they will be able to successfully do an update to
correct the problem.

Change-Id: I02c3894718d595282f57ca63dc2c2ba9b7baa440
Closes-Bug: #1591341