Check policy of resource type again
We allow to delete a stack which is in *_IN_PROGRESS state
and locked by other alive engine. There may be additional
resources that we don't know about if an update was
in-progress when the stack was stopped, so we reload the
stack from the database again before delete the stack, we
have to check the resource policy again in this case.

Change-Id: Id9d6d64f3179a18fedf80449d1e4ec8efae03e4e