Add OS::Heat::Value resource
The OS::Heat::Value resource provides a "value" attribute that may be
calculated through references to other parameters or resources. Other
resources may then reference this value.

The primary motivation is to avoid having to create nested templates
that only exist to execute these kinds of data transformations.

This resource's "type" property is optional but recommended.

Change-Id: I72fdab8261336644a9545668fac6f4e49a785e36