Don't change properties in regenerate_info_schema
The definition of TemplateResource.regenerate_info_schema() has the effect
of changing the semantics of handle_update() in odd ways. For all other
resource types, self.properties is set to the existing properties
definition of the resource at the time that handle_update() is called
(passing the new resource definition as an argument. However,
TemplateResource is an exception because regenerate_info_schema() replaces
self.properties with the new values from the new resource definition.

To reduce confusion and make all resources consistent in their semantics,
this change refactors the change of the properties definition out of the
_generate_schema() method that is called by regenerate_info_schema().

Change-Id: Iab23ae39d05f8290dad02775dddd2e856b29951a
Related-Bug: #1584623