Fix devstack plugin
When heat is enabled, both stack.sh and plugin try to do account
creation (as part of keystone startup which fails as heat is not yet
installed), intialize the db and start the services. In this patch
(which is a workaround), we'll use a set of dummy function calls to
avoid it. We'll remove these functions later(once we clean up devstack
stack.sh) and rename the currently used functions.

Change-Id: I84250d598ba76deae2a5bad6bbf3f0acfe1cfbe6