Do str_replace in a single pass
Previously we implemented the str_replace intrinsic function by repeatedly
calling str.replace(), with the result that some text could be subject to
multiple substitutions (i.e. some new text that was inserted by a
replacement could itself be partly replaced if some part of it matched
another key), although users could not rely on this happening since the
replacement order is explicitly undefined.

We anticipated that users would guard against problems here using some sort
of guard characters like %replace_me% (although we avoided requiring any
particular choice of characters). Over time it has become clear that for a
variety of reasons (of varying legitimacy), this is often not the case.
Sometimes the format of the source data is not under the direct control of
the user.

This patch revises the algorithm to perform the replacements in a single
pass and makes the ordering of keys deterministic. It also refactors the
various versions of the function so that we no longer need two
implementations of the core algorithm.

Change-Id: I43461a0d78226f921018efb32aae6e2255928bd0