Parse JSON as JSON when converting to YAML
The convert_json_to_yaml function is a horror of almost Lovecraftian
propportions, designed to convert from JSON to YAML without changing the
order of keys in the file.

The fix for bug 1286380 (which appears to be resolved in the latest
versions of PyYaml anyway) means that, if this was not corrupting data
before, then it certainly can now.

This patch takes advantage of the fact that we no longer support Python 2.6
to just parse the JSON data as JSON (instead of YAML) using the
object_pairs_hook to ensure that the output order does not change.

Change-Id: I4b09b4c913ec6f8d9c4b7017e3a7a5b7166cf78e
Related-Bug: #1286380
Closes-Bug: #1613881