Fix building derived inputs from a derived config
The previous patch for this bug, bbb1dfb06e074dd6da285723393cea7cf419f9b7,
changed from using the previous software config to calculate the
previous inputs to using the previous derived config. Since the derived
config includes values, this caused an error when trying to initialise
the InputConfig objects in _build_derived_inputs(). That'd be easily
fixed, but it's probably overkill anyway now that we have the entire
derived config, so just use the stored values.

Change-Id: I7424df9a564d63eb197da93b16409149a4f37fdb
Closes-Bug: #1616550