Fix incorrect behaviour of repeat function
If template equals to int, float or boolean value,
repeat function ignores it and returns None, which is
incorrect. This patch fixes such behaviour - now if
template doesn't need a replacement, it's returned without
any changes.

Note that if function looks like:

repeat:
  template: True
  for_each: ['a', 'b', 'c']

result of the function will equals to [True, True, True].

Change-Id: Id7785daa49fe1377b4ce3f9f39ca4b649ea4af18
Closes-bug: #1617203