Allow any data to be passed to yaql function
There's no reason whatsoever that we can't pass data of any type, not
just dicts, to YAQL. It happily accepts lists, strings, ints, &c. Even
if it did not, we are actually passing a dict with a key named 'data'
that contains the actual data, so the data passed to YAQL is always a
dict. There's even less reason to restrict the contents of that first
key to also be a dict.

Since Newton will be the first release to include the yaql function, we
can relax the type checking now without creating any backward
compatibility issues.

This change does that and also simplifies the implementation of the Yaql
function class and stops it from doing dodgy things like overwriting the
expression with its resolved value.

Change-Id: Ia2c699f965c7fb59f74cc36fb1e14200dbbf7e74