Use physical name in event if resource_id is None
An event for CREATE_IN_PROGRESS will have a physical_resource_id value
of None because the resource hasn't been created yet.

Because of this it is not possible to relate a nested stack event with
the CREATE_IN_PROGRESS resource event in the parent stack.

This change populates Event.physical_resource_id with
Resource.physical_resource_name() if Resource.resource_id isn't
populated yet.

This change will help fix
Related-Bug: #1619415

Change-Id: Ib8feb7752bd5736785d142216312bb35629b3601