Get rid of Template.validate_condition_definitions()
Following on from the deprecation of validate_resource_definitions() from
the Template interface, also remove validate_condition_definitions(). The
definitions will be validated when calling Template.conditions(), which
now occurs when loading the resource definitions, so it will still happen
during Stack.validate().

Since this interface is new and has never appeared in a release, we have no
need to go through a deprecation process.

Change-Id: I4dc36433c1aca776c2151205ed26a0f5241a750a