Pre-validate cinder quotas with the real fact
This patch aim to provide some validation to OS::Cinder::Quotas
to check if we accept those quotas base on the real fact we have.
For example, if we have 100 volumes and we set quota `volumes` to 1.
That quota will be pointless (event if can prevent further creation of
volume). With management consideration, this patch add validation before
actually set quota in cinder.
Closes-Bug: #1620226

Change-Id: Ib717c78cd4a220e00ff28fb24d5a8375aa19eb1b