Non-destructive upgrade for structured deployments
If you attempt to update a stack containing
OS::Heat::StructuredDeployments resources, so it uses the new
non-deprecated OS::Heat::StructuredDeploymentGroup type instead,
it deletes the group, and all of the deployments. This means that
any deployment "actions" property will be misinterpreted, e.g if
you have actions: CREATE, all the deployments will re-run on the
update, even though it's an update, not a create.
This patch use substitute class for OS::Heat::StructuredDeployments.
Closes-Bug: #1622056

Change-Id: I7571c4591c32009814a11c71f785b395055e941e