Set hooks before storing events
There is a race condition where setting a hook, where we store the event
before and only then set the hook. If somehow if a test is fast enough
to retrieve the events and signal before the hook is actually set, it
will fail. It makes more sense to set the hook first.

Change-Id: I8f3c33da37419d2d9ead27aaf43949e876236bc5
Closes-Bug: #1626075