Do not attempt deletion of a DELETE_COMPLETE stack in service api
A stack may be in transient state where it is DELETE_COMPLETE, but has
has not actually been soft-deleted yet. For the purposes of
delete_stack in service.py, consider a DELETE_COMPLETE stack as
equivalent to a soft-deleted one (it soon will be), thereby avoiding a
race where we would have attempted to update the stack, running into a
foreign-key constraint issue for a non-existing user_cred.

Change-Id: Iec021e6a0df262d447fdf9ee1789603c7a1c55f8
Closes-Bug: #1626173
Closes-Bug: #1626107