Create a root Yaql context
In the yaql function, we create and store a yaql context object that we
keep during the lifetime of the function. This is only needed for
evaluation. Not keeping track of the context is an improvement on
memory, but require registration of the library every time. The most
efficient way to use yaql contexts seems to be to create a root context,
and then pass a child one for each evaluation.

Change-Id: I12ea701e51a4c39b5a28d4bd4b61a67ab34dc16b
Partial-Bug: #1626675
(cherry picked from commits 0830318707b6b124dc171fdf143fd4117091420f
                        and 8b7e5bee3b190cf83be7b8292fff012c58d970ef)