Properly compute resource_properties pickle
When you check for resource_properties size when storing event, we use
the regular pickle module with the default protocol. Sqlalchemy uses
cPickle if available, with the highest protocol, which can result in
fairly different storing size. Use that instead, this should remove the
last occurence of the storage bug.

Change-Id: Ia417e221f7097a388a6b36959064b12cb818206f
Related-Bug: #1493858