Keep same physical_resource_name in backup stack
When a resource was moved to the backup stack, calls to
Resource.physical_resource_name() would return a different name (with an
asterisk appended to the end of the stack name) than it did previously.
This made it unreliable to use the physical_resource_name() method within
handle_delete(), since it is supposed to return a consistent name.

Change-Id: Ia2699d66135f767fb0bd7b63ec3039fbf283fb38
(cherry picked from commit 7bf55275931a2006fde37d3018929207596df852)