dd missing memory.resident to NOVA_METERS
memory.resident is missing from NOVA_METERS list. This causes the
wrong metadata format for alarms on this meter if a heat stack
is used to create it. Basically, you will see the query in the
alarm generated by the heatstack will incorrect be set to
metadata.metering.stack instead of metadata.user_metadata.stack.

Change-Id: Ic95f8c98c020627c6c5322c905cbf1c160d8445b
Closes-Bug: #1623680