Avoid redundant validations
When multiple resources in a stack have identical definitions, there is no
purpose in re-running validation on every single one. This is a common
occurrence in, for example, autoscaling groups.

By eliminating the need to validate every identical resource, this patch
should help prevent timeouts on create/update with large autoscaling groups
of the kind created by e.g. Magnum.

Change-Id: Iaed5a707a72ae558552d35738db10dc05e5f6343
Closes-Bug: #1644013