Don't return the signed queue URL from get_resource
A signed Zaqar queue URL is too long to fit as a physical resource ID.
The current queue URL is also very specific to Aodh, and likely won't
work anywhere else.

To solve both of those issues, no longer return the URL from
get_resource and instead add a "query_str" attribute, which can be used
both for Aodh and likely other places. Also add a "project" attribute
and store it in the resource data.

Change-Id: Ib5a4091c1bcd39f113843ac60022307917a07804
Closes-Bug: #1645888