Use BaseServer base class for Nova Server resource
Refactors the Nova Server resource class to use a new base class,
BaseServer.  All of the Nova/Neutron/Glance/etc specific logic for the
Server resource remains in the Server class and not in the base class.

This allows other Server subclasses, such as the DeployedServer resource
(see followup patch) to create Server "like" resources to associate with
SoftwareDeployment resources.

Partially-implements: blueprint split-stack-software-configuration
Change-Id: I585d92dbd29198107464b92ddb0e0b15779b9999