map_replace allow noop colliding replacements
We should allow the case where the keys collide, but the
replacement is the same as the current key (e.g a noop).

This is useful when you want to do a map_replace based on a
user parameter, then the default can be to do a noop replace
but the user parameter may also be any other (non colliding)
value.

Change-Id: I8cc5761c219616b4f8c18b3f44c4bf864c5457f1
Closes-Bug: #1652034