Efficient passing of attrs during traversals
Two improvements:

Do not iterate through stack outputs when determining what attributes
to send as input_data to the next resources in a traversal; it is at
best extra processing and at worse results in extra attributes being
included in input_data.

Do not re-resolve attributes / re-calculate input_data when we already
have it. I.e., when a resource constructs input_data to send to a
requirer resource, that same input_data may be used for all other
requirer resources.

Change-Id: I64089fb0774c10f172d986c3f87090e91cb3f263
Closes-Bug: #1656125