Ignore dependency errors with conditions
If a resource is conditionally defined, depending on it raises an error
in the case it's not defined. This patch fixes that by checking if the
resource is present in the template regardless of the conditions.

Change-Id: Iefae1fcea720bee4ed69ad1a5fe403d52d54433c
Closes-Bug: #1649900
(cherry picked from commit e1aeabaa032c607b07aa835526c06aad0d1e21c5)