Use stored value in OS::Heat::Value resource
Currently when getting the 'value' attribute of an OS::Heat::Value
resource, we recalculate the value from the properties - which means
re-resolving any functions, including get_attr and the like. That's
expensive and opens the possibility that the value can change over time,
which is probably not what users would expect.

Use the property values stored in the database at the time the resource was
created or last updated instead.

Change-Id: If0f80cab94c28514d1569b1025362ab9d9d31512
Closes-Bug: #1661728