Still wait for hooks on failed resources
Previously, if a resource was in a FAILED state Heat would set the hook
but not wait for the user to clear it before continuing. It is not clear
what case this was designed to handle, since there's no way that I know
of for the resource to go into a FAILED state while waiting for the
hook.

Change-Id: Iaacc585b9806c08fd5a57c80c10eeb31cb0a8f71
Closes-Bug: #1665699