Add a StackDefinition class
This defines clearly the API through which Template/Function plugins
may access data about the stack. That allows us to ensure that we don't
accidentally access data in a highly inefficient manner, and third-party
plugin developers to have confidence that the interface they're using is
stable and won't change without a multi-cycle deprecation process with
ample warning.

Change-Id: I7a67700685770d6d5b9c811e72a8b21221400092
Partially-Implements: blueprint stack-definition