To reset metadata for resources when mark healthy
We reset the metadata for resources when mark healthy
in change Ibd6c18acf6f3f24cf9bf16a524127850968062bc,
but it doesn't work, because resource policy and group
inherit from multiple class, and the method
'handle_metadata_reset' actually be called now is wrong.

This patch will fix it.

Change-Id: I8583f26b6c8f6c814a00a0a38042484378fa53db
Closes-Bug: #1672596