Delete software deployment before credentials
When deleting a software deployment, remove the deployment from the server
first (by deleting the derived config) and only *then* delete the user
credentials that the server might use to signal back to Heat. This reduces
the chance of something failing that leaves things at a point where revoked
credentials on an old deployment are blocking os-collect-config.

Change-Id: If5fe38033a8a402b07266b1c2d4b627efff85d6b