Copy template version when update fails
When an update fails, we may have copy some chunk of resources or
parameters to the new template. If the version was updated and the new
resources require the version, this can lead to a state where the stack
is in an usable state. This synchronizes the version when a failure
happens.

Change-Id: I2faf8f3541fc800ea61c417e5575f4a56a83665b
Closes-Bug: #1620696
(cherry picked from commit 45fde101979d9d3f446cd937e5cc4c47539cc5bc)