Fix handle_restore for server and volume resources
When resource references are used for properties restore
does not work for nova server and volume resources.

This patch fixes it and adds a functional test for it.

 Conflicts:
	heat/engine/resources/openstack/nova/server.py

There is very minor conflict in server.py as the code
has been changed in pike for translation refactoring.

Change-Id: I34efb3f4e0d3578d95134793f5c5d413667790e0
Closes-Bug: #1687009
(cherry picked from commit a8129573d221284a482e84e62a7e8b0732318bfe)