Always return live resource attributes in list_stack_resources
Currently, we do live resolution of attribute values when in
list_stack_resources with detail enabled. However, in future we will cache
in the database those values that are referenced elsewhere in the template.
Since it's unlikely that every attribute would be referenced in the
template, this could lead to the API returning an inconsistent mix of live
and cached data.

Caching all of the attribute values on the off chance that someone might
request them through this obscure API would be expensive, and there are
also users that need to be able to access live resource attributes via
Heat. Therefore, always return live data for the resource attributes in
this API.

Change-Id: Ifea71b30299076d14d9bb91f3bce4dd7fc4c88a1
Related-Bug: #1660831