Document user_data replace-on-update policy
Since d4188127a14686f6d9844180b977cf5fa05aa024 the documentation for the
user_data property of OS::Nova::Server has been dangerously misleading,
stating that updates to the property don't cause replacement of the
resource, which is not true by default. This adds documentation that
explains exactly what is going on.

Change-Id: I2d95205fbcbd0bfafa6690a6efec1a56441ce22d