Mark Server os_collect_config attribute as CACHE_NONE
When the os_collect_config attribute was added to OS::Nova::Server (in
64ddd071604bf4fbdeb97f8c8fa4e313dd611c3c), it was implemented both in
_resolve_attribute() and in get_attribute(), with the latter overriding the
former so it is never called (for this attribute) and so that the value is
never cached. To be sure, to be sure, the attribute caches - which cannot
contain the value of this attribute - were cleared whenever the deployment
data is updated.

It's much easier to just mark the attribute as CACHE_NONE. This also makes
the status of this attribute visible to other parts of the code.

Change-Id: Icaecf70529f19fb32b34de9cedb82e5006282002
Partial-Bug: #1660831
Partially-Implements: blueprint split-stack-default