Fix nonsensical docs in dependencies.py
It's really confusing every time you have to think about the difference
between requires() and require_by() in the dependencies graph. That was
helped not at all by 5189bbebabbbe2c672c8bb492c496e5780892d01, which
added a requires() method to the Dependencies class with a parameter
name that is exactly backwards from correctly representing what it
means, as well as a docstring that is copied without modification from
the required_by() method, which obviously returns the opposite data. In
addition, the new docstring for the Node.requires() function made no
sense.

Change-Id: Ie9d8bedf83947b197f2defd309fd996060514100